requirements.txt
aws-cdk.core==1.82.0
aws-cdk.aws_ec2==1.82.0
aws-cdk.aws_ecs==1.82.0
aws-cdk.aws_ecs_patterns==1.82.0